test: 인수테스트 부모 클래스를 추가하라, 테스트에서 공통으로 쓰이는 것에 책임은 부모클래스에서 갖는다
