+++ src/test/java/nextstep/subway/station/StationAcceptanceTest.java
@@ -3,14 +3,10 @@ package nextstep.subway.station;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.subway.db.DataInitializer;
+import nextstep.subway.AcceptanceTest;
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
@@ -25,22 +21,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
 @DisplayName("지하철역 관련 기능")
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class StationAcceptanceTest {
-    @LocalServerPort
-    int port;
-
-    @Autowired
-    DataInitializer dataInitializer;
-
-    @BeforeEach
-    void setUp() {
-        RestAssured.port = port;
-    }
+class StationAcceptanceTest extends AcceptanceTest {
 
     @AfterEach
     void cleanUp() {
-        dataInitializer.execute("station");
+        databaseClean("station");
     }
 
     /**
